--- cross_diffusion/bifurcation_surface/.signac_shell_history
@@ -1,9 +1,0 @@
-help(project)
-job
-print(job)
-job = project.get_job()
-exit()
-with signac.H5Store('shared_data.h5').open(mode='r+'):
-with signac.H5Store('shared_data.h5').open(mode='r+') as sd:
-	sd['N_n'] = 200
-exit()
