--- cross_diffusion/bifurcation_surface/signac.rc
@@ -1,5 +1,0 @@
-project = None
-schema_version = 1
-workspace_dir = workspace
-[General]
-[hosts]
