--- generalized_models/signac.rc
@@ -1,2 +1,0 @@
-project = None
-schema_version = 1
