--- generalized_models/signac_project_document.json
@@ -1,1 +1,0 @@
-{"_status": {}}
